Changed class name and removed unnecessary codes
